+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Internal changes & Bugfixes**
 * [HELM] Added option to specify image pull secret name.
 * Fixed #1948: Sanitise user input in error output.
+* Fixed MultiDatastream not picking up ID type of Thing and Sensor.
 
 
 ## Release version 2.3.2